printbackend/backend/requirements.txt
django>=5.0
djangorestframework
psycopg2-binary
django-cors-headers
whitenoise
playwright
Pillow
gunicorn
dj-database-url
python-dotenv
drf-nested-routers
requests
reportlab
razorpay
boto3
django-storages
